requirements.txt
anyio==3.5.0
asgiref==3.5.0
astroid==2.9.3
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11; python_version >= '3'
click==8.0.3
django-environ==0.8.1
fastapi~=0.74.1
google-api-core==2.5.0
google-api-python-client~=2.37.0
google-auth-httplib2==0.1.0
google-auth==2.6.0
google==3.0.0
googleapis-common-protos==1.54.0
greenlet==1.1.2; python_version >= '3' and (platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))))
h11==0.13.0
httplib2==0.20.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
oauth2client==4.1.3
platformdirs==2.4.1
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydantic==1.9.0
pyjwt==2.3.0
pylint==2.12.2
pyparsing==3.0.7; python_version > '3.0'
requests==2.27.1
rsa==4.8; python_version >= '3.6'
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlalchemy==1.4.29
starlette==0.17.1
toml==0.10.2
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.4
wrapt==1.13.3

passlib~=1.7.4
jwt~=1.3.1